requirements.txt
discord.py
psycopg2-binary
python-dotenv
openai
matplotlib
google-auth
google-auth-oauthlib
google-api-python-client
Flask
pytz